refactor: streamline utility imports
